refactor(privacy_preserving_circuit): extract execution_state module
Refs: #454
